refactor requestid
